+++ test/terra/backends/test_qasm_statevector_simulator.py
@@ -47,6 +47,8 @@ from test.terra.backends.qasm_simulator.qasm_snapshot import QasmSnapshotStateve
 from test.terra.backends.qasm_simulator.qasm_snapshot import QasmSnapshotDensityMatrixTests
 from test.terra.backends.qasm_simulator.qasm_snapshot import QasmSnapshotStabilizerTests
 from test.terra.backends.qasm_simulator.qasm_snapshot import QasmSnapshotProbabilitiesTests
+from test.terra.backends.qasm_simulator.qasm_snapshot import QasmSnapshotExpValPauliTests
+from test.terra.backends.qasm_simulator.qasm_snapshot import QasmSnapshotExpValMatrixTests
 # Other tests
 from test.terra.backends.qasm_simulator.qasm_method import QasmMethodTests
 
@@ -77,6 +79,8 @@ class TestQasmStatevectorSimulator(common.QiskitAerTestCase,
                                    QasmSnapshotStatevectorTests,
                                    QasmSnapshotDensityMatrixTests,
                                    QasmSnapshotProbabilitiesTests,
+                                   QasmSnapshotExpValPauliTests,
+                                   QasmSnapshotExpValMatrixTests,
                                    QasmSnapshotStabilizerTests):
     """QasmSimulator statevector method tests."""